added static cast in make decision

+++ src/BotPlayer/BotPlayer.cpp
@@ -47,7 +47,7 @@ void BotPlayer::make_decision(unsigned int money_to_bet, int num_of_players){
     if (money_to_bet == 0 && equity() < 0.4) {
         make_check();
     } else if (equity() > 1.0 / static_cast<double>(num_of_players)) {
-        (money() <= money_to_bet) ? make_all_in() : make_raise(money_to_bet + (money() - money_to_bet) * equity());
+        (money() <= money_to_bet) ? make_all_in() : make_raise(money_to_bet + static_cast<int>((money() - money_to_bet) * equity()));
     } else if (equity() < 0.5 / static_cast<double>(num_of_players)) {
         make_fold();
     } else {
